refactor: move data model classes to the model package
