+++ app/src/main/java/campus/tech/kakao/map/model/HistoryItem.kt
@@ -1,3 +1,3 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.model
 
 data class HistoryItem(val name: String)
