+++ app/src/main/java/campus/tech/kakao/map/model/PlaceItem.kt
@@ -1,3 +1,3 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.model
 
 data class PlaceItem(val name: String, val address: String, val category: String)
